+++ src/CMP_demo/user_interface/components/UI/Button/Button-styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+export const Btn = styled.a`
+  background-color: #00518f;
+  color: #FFF;
+  padding: .6rem 1.2rem;
+  border-radius: 10px;
+  transition: background-color .4s;
+  cursor: pointer;
+
+  &:hover {
+    background-color: #4b7391;
+  }
+`
