+++ src/CMP_demo/user_interface/components/UI/Button/Button.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import { Btn } from './Button-styles'
+
+const button = (props) => (
+  <Btn onClick={props.clicked} >{props.label}</Btn>
+)
+
+export default button
