+++ src/CMP_demo/user_interface/components/UI/Code/Code-styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const CodeBox = styled.pre`
+  width: 100%;
+  background-color: #EEEEEE;
+  padding: 1rem;
+`
