+++ src/CMP_demo/user_interface/components/UI/Code/Code.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { CodeBox } from './Code-styles';
+
+const code = (props) => (
+  <CodeBox>
+    <code>
+{`${props.code}`}
+    </code>
+  </CodeBox>
+)
+
+export default code
