+++ src/CMP_demo/user_interface/components/UI/Nav/Nav-styles.js
@@ -1,0 +1,81 @@
+import styled from 'styled-components'
+
+export const Navigation = styled.div`
+  position: fixed;
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  padding-left: 2rem;
+  padding-top: 2rem;
+  width: 18.75rem;
+  background-color: #00518F;
+
+  a {
+    position: relative;
+    color: #FFFFFF;
+    text-decoration: none;
+
+    &:after {
+      position: absolute;
+      bottom: -7px;
+      left: 0;
+      content: '';
+      background-color: #b7f0ff;
+      height: 2px;
+      width: 0;
+      transition: width .3s, height 1s;
+    }
+
+    &:not(:last-child) {
+      margin-bottom: 1.2rem;
+    }
+
+    &:hover {
+      &:after {
+        width: 200px;
+      }
+    }
+    
+    &:first-child {
+      font-weight: bold;
+      font-size: 1.6em;
+      color: #b7f0ff;
+      margin-bottom: 1rem;
+
+      &:after {
+        content: none;
+      }
+    }
+  }
+
+  .active {
+    color: #b7f0ff;
+    &:after {
+      position: absolute;
+      bottom: -7px;
+      left: 0;
+      content: '';
+      background-color: #b7f0ff;
+      height: 2px;
+      width: 0 !important;
+      transition: width .3s;
+    }
+  }
+`
+
+export const Subtitle = styled.div`
+  position: relative;
+  text-transform: uppercase;
+  margin-bottom: 1rem;
+  margin-top: 1rem;
+  color: #FFF;
+
+  &:after {
+    content: '';
+    position: absolute;
+    left: 0;
+    bottom: -5px;
+    width: 30px;
+    height: 2px;
+  }
+`
