+++ src/CMP_demo/user_interface/components/UI/Output/Output-styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const CodeBox = styled.pre`
+  border: 1px solid #BBBBBB;
+  display: block;
+  max-width: calc(100vw - 22.75rem);
+  padding: 1rem;
+  overflow: scroll;
+`
