+++ src/CMP_demo/user_interface/components/UI/Output/Output.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { CodeBox } from './Output-styles'
+
+const code = (props) => (
+  <CodeBox>
+    <code>
+{`${props.output}`}
+    </code>
+  </CodeBox>
+)
+
+export default code
