+++ src/CMP_demo/user_interface/components/UI/PageMainTitle/PageMainTitle-styles.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const Title = styled.h2`
+  color: #00518F;
+`
