+++ src/CMP_demo/user_interface/components/UI/PageMainTitle/PageMainTitle.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import { Title } from './PageMainTitle-styles'
+
+const pageTitle = (props) => (
+  <div>
+    <Title>{props.title}</Title>
+  </div>
+)
+
+export default pageTitle
