+++ src/CMP_demo/user_interface/index.js
@@ -1,0 +1,13 @@
+import React from "react"
+import ReactDOM from "react-dom"
+import { BrowserRouter } from 'react-router-dom'
+
+import App from "./components/App"
+
+import './index.css'
+
+ReactDOM.render((
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+), document.getElementById('root'));
