+++ src/CMP_demo/user_interface/pages/Home/Home.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const home = () => (
+  <div>
+    <p>Página de Demo de Schibsted CMP</p>
+  </div>
+)
+
+export default home
