+++ src/CMP_demo/user_interface/pages/MethodPages-styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const Subtitle = styled.h3`
+  margin-top: 1.5rem;
+  margin-bottom: 1.5rem;
+  font-weight: 400;
+`
+
+export const WrapperButtons = styled.div`
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+
+  a:first-child {
+    margin-right: 1.5rem;
+  }
+`
