--- src/index.js
@@ -1,13 +1,0 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import { BrowserRouter } from 'react-router-dom'
-
-import App from "./components/containers/App"
-
-import './index.css'
-
-ReactDOM.render((
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>
-), document.getElementById('root'));
