--- src/pages/Home/Home.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const home = () => (
-  <div>
-    <p>Minions ipsum baboiii chasy wiiiii hana dul sae daa aaaaaah me want bananaaa! Tulaliloo tank yuuu! Poopayee me want bananaaa! Bappleees. Hahaha la bodaaa bee do bee do bee do me want bananaaa! Gelatooo bee do bee do bee do poulet tikka masala baboiii underweaaar tatata bala tu wiiiii wiiiii para tú. Daa tulaliloo bee do bee do bee do gelatooo me want bananaaa! La bodaaa tulaliloo potatoooo pepete daa daa. Jeje bee do bee do bee do butt aaaaaah gelatooo po kass bee do bee do bee do tatata bala tu jeje. Bappleees jeje bananaaaa po kass para tú. Bappleees butt ti aamoo! Aaaaaah bappleees bee do bee do bee do.</p>
-    <p>Uuuhhh hahaha bappleees me want bananaaa! Bappleees. Bee do bee do bee do tank yuuu! Baboiii tatata bala tu baboiii tank yuuu! La bodaaa. Underweaaar la bodaaa daa belloo! Daa po kass butt poulet tikka masala. Belloo! butt chasy bananaaaa aaaaaah po kass jiji.</p>
-    <p>Hahaha pepete daa bappleees tank yuuu! Wiiiii pepete pepete. Jeje bananaaaa hahaha bananaaaa tank yuuu! Potatoooo pepete tulaliloo poopayee poulet tikka masala para tú. Potatoooo hana dul sae potatoooo aaaaaah. Gelatooo poulet tikka masala wiiiii poopayee me want bananaaa! Poulet tikka masala tulaliloo tulaliloo. Uuuhhh uuuhhh bananaaaa tank yuuu! Underweaaar poopayee daa tank yuuu! Tatata bala tu bappleees jiji aaaaaah potatoooo po kass para tú hana dul sae uuuhhh pepete.</p>
-    <p>Jiji poulet tikka masala underweaaar bee do bee do bee do uuuhhh tank yuuu! Bananaaaa hahaha wiiiii bappleees underweaaar belloo! Baboiii gelatooo gelatooo butt. Tulaliloo daa bananaaaa underweaaar tank yuuu! Tank yuuu! Jiji la bodaaa butt chasy. Gelatooo la bodaaa bananaaaa para tú gelatooo bappleees gelatooo potatoooo potatoooo aaaaaah. Poopayee baboiii tank yuuu! Baboiii la bodaaa aaaaaah gelatooo po kass wiiiii butt bappleees. Tank yuuu! hahaha po kass belloo!</p>
-  </div>
-)
-
-export default home
